Add key/button for activating movement
If activate_is_toggle is true, then the configured mouse button and/or
key will toggle movement on/off. If activate_is_toggle is false, then
the button must be held down to activate the fly camera movement
systems. This allows for example holding down the right mouse button for
navigating a scene or releasing it to interact with UI elements.